refactor: 중복 디렉토리 삭제 (#10)
